requirements.txt
streamlit>=1.21.0
altair<5
pytest
anthropic>=0.39.0
python-dotenv>=1.0.0
